Atualização

+++ site/src/database/script-tabelas-sqlServer.sql
@@ -107,11 +107,11 @@ CREATE TABLE Usuario (
 CREATE TABLE Parametrizacao (
 	id INT NOT NULL IDENTITY(1, 1),
 	empresa_id INT NOT NULL,
-	qtd_cpu_max INT NULL DEFAULT 90,
-	qtd_bytes_enviado_max INT NULL DEFAULT 100000,
-	qtd_bytes_recebido_max INT NULL DEFAULT 100000,
-	qtd_memoria_max INT NULL DEFAULT 90,
-	qtd_disco_max INT NULL DEFAULT 90,
+	qtd_cpu_max BIGINT NULL DEFAULT 90,
+	qtd_bytes_enviado_max BIGINT NULL DEFAULT 100000,
+	qtd_bytes_recebido_max BIGINT NULL DEFAULT 100000,
+	qtd_memoria_max BIGINT NULL DEFAULT 90,
+	qtd_disco_max BIGINT NULL DEFAULT 90,
 	PRIMARY KEY (id),
 	FOREIGN KEY (empresa_id) REFERENCES Empresa (id) ON UPDATE CASCADE ON DELETE CASCADE
 );
